Ratings Service (#315)
* feat(ratings-app): add ItemPager interface and implementation to RatingsService to improve code reusability and readability
feat(ratings-app): add ItemPager interface and implementation to ItemPager.kt to improve code reusability and readability
refactor(ratings-app): remove unused import and variable in RatingsPresenter to improve code readability
refactor(ratings-app): simplify RatingsService message generation and add NOTION_URL constant to improve code readability and maintainability
refactor(ratings-app): simplify RatingsService constructor to improve code readability and maintainability

* refactor(AuthServer.kt): make notionHttpClient public to be able to use it in other modules
refactor(AuthServer.kt): remove onBrowse parameter from getAccessToken function as it is not used anymore
refactor(AuthCommand.kt): remove onBrowse parameter from AuthCommand function as it is not used anymore
refactor(AuthCommand.kt): remove userPromptUri parameter from AuthState.Awaiting as it is not used anymore
feat(AuthCommand.kt): get access token synchronously instead of asynchronously to simplify code and improve readability

* refactor(AuthServer.kt): rename Notion.Error to Notion.Object.Error to improve semantics
feat(AuthServer.kt): add support for saving access token to a file to avoid requesting a new token on every app start
feat(AuthServer.kt): add support for loading access token from a file to avoid requesting a new token on every app start
feat(AuthServer.kt): add support for Notion.Object.Search to be able to search for objects in Notion API

* feat(notion): add support for Notion API integration to fetch ratings items from a database and rate them based on ELO algorithm
refactor(ratings): change score type from Long to Double to support fractional scores
refactor(ratings): use Notion HTTP client instead of local HTTP client to fetch ratings items from Notion database
refactor(ratings): change RatingsService implementation to use ItemGenerator to fetch ratings items from Notion database

* refactor(RatingsScreen.kt): replace Column with Box to simplify layout and improve performance
style(RatingsScreen.kt): remove unused imports and commented out code

* refactor(RatingsScreen.kt): reduce card padding from 12dp to 4dp and add 12dp padding to the box inside the card
feat(RatingsService.kt): add support for updating page score in Notion API when rating is updated. Also, add support for fetching database ID by name and fetching all pages in a database.

* refactor(AuthServer.kt): add braces to if statement for better readability and maintainability

* style(RatingsScreen.kt): increase padding from 12dp to 24dp for better UI spacing
feat(RatingsService.kt): add support for updating page score using PATCH request instead of POST request for better RESTful API semantics. Also, update the default score value to 1500.

* refactor(AuthServer.kt, AuthCommand.kt, RatingsModule.kt): extract openUri parameter to allow for dependency injection and improve testability

* refactor: remove unused notionHttpClient variable from AuthServer.kt
refactor: remove unused loading extension property from RatingsPresenter.kt
fix: set correct value parameter in updatePageScore function in RatingsService.kt

* refactor(RatingsService.kt): extract ANSI color codes to constants for better readability
feat(RatingsService.kt): add rounding function to round doubles to a specified scale
feat(RatingsService.kt): improve output of rate function to include more information about the rating calculation
feat(RatingsService.kt): improve output of update function to include an arrow indicating the direction of the change in score

* refactor(RatingsService.kt): remove print statements and simplify code by removing unnecessary join() calls

---------

+++ ratings-app/src/commonMain/kotlin/io/ashdavies/playground/RatingsItem.kt
@@ -4,6 +4,6 @@ internal data class RatingsItem(
     val id: String,
     val name: String,
     val ignored: Boolean,
-    val score: Long,
+    val score: Double,
     val url: String,
 )
